Fix structure to actual in master branch. Fix pytest config
