+++ .vscode/settings.json
@@ -7,4 +7,7 @@
     "code-block-style": false
   },
   "markdown.validate.enabled": true,
+  "python.analysis.extraPaths": [
+    "./src"
+  ],
 }
