+++ pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+pythonpath = src
+
+testpaths =
+    tests/pytest
